pyproject.toml
[tool.black]
line-length = 98

[tool.poetry]
name = "presentation-mode"
version = "0.1.0"
description = ""
authors = ["Simon Brulhart <simon@brulhart.me>"]

[tool.poetry.dependencies]
python = "^3.7"
DBussy = "^1.2"
slackclient = "^2.1"
python-decouple = "^3.1"

[tool.poetry.dev-dependencies]
ipdb = "^0.12.2"
black = {version = "^18.3-alpha.0", allows-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
